% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_diagnostics_mcmc_output.R
\name{analyse_mcmc_convergence2}
\alias{analyse_mcmc_convergence2}
\title{Dispays output diagnostics}
\usage{
analyse_mcmc_convergence2(
  mcmc_sims,
  states,
  par_names,
  par_names_plots,
  lab_names = NULL,
  start_vals,
  true_vals = NULL,
  burn,
  thin = NULL,
  ki_prob = 0.9,
  plot_view = FALSE,
  plot_ggp2 = FALSE,
  plot_save = FALSE,
  plot_name = "",
  plot_path = NULL,
  table_view = FALSE,
  table_save = FALSE,
  table_name = "",
  table_path = NULL,
  table_prec = 4,
  compute_ess = TRUE,
  compute_ess_stan = FALSE,
  ur_view = FALSE,
  ur_save = FALSE,
  ur_name = "",
  ur_path = NULL
)
}
\arguments{
\item{mcmc_sims}{simulated draws/ mcmc output: matrix of dimension "simulated
MCMC draws" (rows) x "parameters" (cols)}

\item{states}{state trajectories i.e. particle filter (SMC) output}

\item{par_names}{parameter names as used in the pgas output (col names for
\code{mcmcm_sims} e.g.)}

\item{par_names_plots}{parameter names for plot labelling}

\item{lab_names}{names for labels}

\item{start_vals}{start values of the parameters}

\item{true_vals}{true values if a simulation is run; default is NULL and true
values will not be added for the histogram and trace plots}

\item{burn}{burnin period}

\item{thin}{thinning; \code{thin = 10} means every 10th draw is excluded}

\item{ki_prob}{probablitlity with which confidence intervals and highest
posterior density regions are computed}

\item{plot_view}{logical; if \code{TRUE}, plots are returned for viewing
(RStudio pane)}

\item{plot_ggp2}{logical; if \code{TRUE},ggplots are generated (if
\code{FALSE}, then base plots are used)}

\item{plot_save}{logical; if \code{TRUE}, plots are saved with names
specified under \code{plot_name} and path given by the argument
\code{plot_path}}

\item{plot_name}{if \code{plot_save = TRUE}, defines the names of the plots
as a character vector}

\item{plot_path}{if \code{plot_save = TRUE}, defines the path where plots are
stored as a character}

\item{table_view}{logical; if \code{TRUE}, output table can be viewd (RStudio
pane)}

\item{table_save}{logical; if \code{TRUE}, output tables are saved}

\item{table_name}{if \code{table_view = TRUE}, defines the name of the output
table as a character}

\item{table_path}{if \code{table_view = TRUE}, defines the path where output
table is stored as character vector}

\item{table_prec}{if \code{table_view = TRUE}, rounding digits for values of
the output table}

\item{compute_ess}{logical; if \code{TRUE}, computes the effective sample
size}

\item{compute_ess_stan}{logical; if \code{TRUE}, computes the effective
sample sizes in terms of ESS bulk and ESS tail (see the function
documentation of \code{diagnostics_table()} for details)}

\item{ur_view}{logical; if \code{TRUE} are returned for viewing (RStuido
pane)}

\item{ur_save}{logical; if \code{TRUE} update rates are saved}

\item{ur_name}{if \code{ur_save == TRUE}, specifies name of the update rate
plot}

\item{ur_path}{if \code{ur_save == TRUE}, specifies path of the update rate
plot}
}
\value{
nothing but plots and table are saved and displayed upon request
}
\description{
Displays output diagnostic plots and tables of (particle) MCMC parameter
draws: histogram, trace and autocorrelations plots for all parameters as
(ggplot2 or base) plots, and an output table summary with posterior mean,
standard deviation of the parameter under the posterior distribution and
standard deviation of the posterior mean, confidence bands, HPDs, and
effective sample sizes (the latter also in the stan variant i.e. ESS bulk and
tail). Also, update rates of simulated latent state parameters (bootsrap
particle filter/SMC output) are returned. All diagnostics can be saved
specifiying a directory, name and if a logical value, \code{..._save},
idnicating whether plots, output diagnostics table or update rates should be
save.
}
