% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_diagnostics_general.R
\name{analyse_states_ur}
\alias{analyse_states_ur}
\title{Computation of update rates for particle states}
\usage{
analyse_states_ur(trajectories, states_in_cols = TRUE, return_values = FALSE)
}
\arguments{
\item{trajectories}{latent state trajectories (with ordering of mcmc draws
and time dimension specified under states_in_cols)}

\item{states_in_cols}{logical with default TRUE; if TRUE, mcmc draws are
given in rows and time dimension (i.e. number of states) in columns}

\item{return_values}{logical; if TRUE, the update rates are returned as
values rather than a plot}
}
\value{
object of type matplot displaying the udpate rates of particle
trajectories
}
\description{
Computes the update rates for simulated/drawn particle trajectories of the
particle MCMC output. Update rates are a measure of how frequently the
particular state (as displayed along the x axes) gets updated in the MCMC
draws of the particle Gibbs/MH approach. The more frequently, the better as
higher update rates indicate a better mixing of the sampler.
}
