% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_diagnostics_general.R
\name{burn_and_thin}
\alias{burn_and_thin}
\title{Eliminate burnin and apply thinning to parameter draws of MCMC output}
\usage{
burn_and_thin(draws, burnin, thin = NULL)
}
\arguments{
\item{draws}{parameter draws of the (particle) MCMC approach in matrix form
where the number of draws are stored in rows and the different parameters
in columns}

\item{burnin}{burnin period (e.g. 1000 = the first 1000 mcmc draws are
eliminated from the chain)}

\item{thin}{thinning frequency (e.g. 10 = every 10th mcmc draw is eliminated
from the chain)}
}
\value{
a matrix of (particle) MCMC parameter draws after burnin and thinning
}
\description{
Given a matrix of mcmc draws, returns a matrix with draws after burnin period
and thinning.
}
