% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{compute_ci}
\alias{compute_ci}
\title{Computes confidence interval}
\usage{
compute_ci(mcmc_sims_after, ki_prob)
}
\arguments{
\item{mcmc_sims_after}{(particle) MCMC draws/simulated samples after burn-in}

\item{ki_prob}{probability mass to cover with confidence bands and HPD}
}
\value{
a two column matrix with upper/lower (first/second column) bound of
a confidence interval with significance given by \code{ki_prob}
}
\description{
Confidence interval is based on samples after burn-in.
}
