% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{compute_hpd}
\alias{compute_hpd}
\title{Compute HPD interval}
\usage{
compute_hpd(mcmc_sims, mcmc_sims_after, burn, num_mcmc, ki_prob)
}
\arguments{
\item{mcmc_sims}{(particle) MCMC draws/simulated samples}

\item{mcmc_sims_after}{(particle) MCMC draws/simulated samples after burn-in}

\item{burn}{burn-in in period}

\item{num_mcmc}{number of MCMC iterations}

\item{ki_prob}{probability mass to cover with confidence bands and HPD}
}
\value{
a two-column matrix that contains the smallest HPD interval for each
parameter
}
\description{
Computes before and after burn-in, and then chooses the smallest of those.
}
