% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{compute_sds}
\alias{compute_sds}
\title{Computes posterior standard deviation & standard deviation of posterior mean}
\usage{
compute_sds(mcmc_sims_after, ess)
}
\arguments{
\item{mcmc_sims_after}{(particle) MCMC draws/simulated samples after burn-in}

\item{ess}{expected sample sizes as produced by \link{compute_ess}}
}
\value{
a two column matrix of two types of standard deviations:
\itemize{
\item posterior standard deviation
\item estimate of standard deviation of the posterior mean based on the
expected sample size
}
}
\description{
Standard deviation of the posterior mean is based on the expected sample
size.
}
