% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{compute_significance_indicator}
\alias{compute_significance_indicator}
\title{Check whether CI and HPD contain true values or zeros.}
\usage{
compute_significance_indicator(int, true_vals)
}
\arguments{
\item{int}{a matrix of two columns with upper/lower confidence bands as
produced by \link{compute_ci} or \link{compute_hpd}}

\item{true_vals}{true values with default set to \code{NULL}}
}
\value{
a list of two vectors giving \code{TRUE} if \code{true_vals} is
covered (or zero, whenever true values are not passed), or \code{FALSE},
if that is not the case for both CI and HPD
}
\description{
The former case is for simulations (with \code{true_vals} being not
\code{NULL}) and the latter checks for significance i.e. if 0 is covered or
not.
}
