% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{diagnostics_table}
\alias{diagnostics_table}
\title{Returns output diagnostics table}
\usage{
diagnostics_table(
  num_par,
  par_names,
  mcmc_sims,
  mcmc_sims_after,
  burn,
  num_mcmc,
  posterior_means,
  start_vals,
  true_vals = NULL,
  ki_prob,
  compute_ess = TRUE,
  compute_ess_stan = TRUE,
  settings_table
)
}
\arguments{
\item{num_par}{number of total parameters (rows of the output table)}

\item{par_names}{names of parameters to be taken as rownames in output table}

\item{mcmc_sims}{(particle) MCMC draws/simulated samples}

\item{mcmc_sims_after}{(particle) MCMC draws/simulated samples after burnin}

\item{burn}{burnin in period}

\item{num_mcmc}{number of MCMC iterations}

\item{posterior_means}{pre-computed posterior means}

\item{start_vals}{starting values of the parameter draws}

\item{true_vals}{true values with default set to \code{NULL}}

\item{ki_prob}{probability mass to cover with confidence bands and HPD}

\item{compute_ess}{logical; if \code{TRUE} computes the effective sample size}

\item{compute_ess_stan}{logical; if \code{TRUE} computes the effective sample
size in stan-style i.e. with ESS bulk and ESS tail}

\item{settings_table}{settings for table output}
}
\value{
returns output diagnostic table with at least 8 columns (mean, sd,
sd-posterior mean, confidence intervall, HPDs) but up to 11 columns (if
ESS, ESS bulk and ESS tail are included)
}
\description{
The output diagnostics table summarizes posterior mean, standard deviation of
the parameter under the posterior distribution and standard deviation of the
posterior mean, confidence bands, HPDs, effective sample sizes (also as stan
version with ESS bulk/tail) in columns for each parameter (in rows).
}
