% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_diagnostics_general.R
\name{equal_values}
\alias{equal_values}
\title{To be used within \code{Reduce()}-constructs to check if values are equal}
\usage{
equal_values(x, y)
}
\arguments{
\item{x}{first value}

\item{y}{second value (that is checked to be equal to \code{x})}
}
\value{
either \code{TRUE} if \code{x} and \code{y} are equal or \code{FALSE}
otherwise
}
\description{
Equal values as returned by \code{all.equal()}, not \code{identical()}.
}
