% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_signal_noise_infos.R
\name{get_signal_to_noise_infos}
\alias{get_signal_to_noise_infos}
\title{Information for diagnostics on regressors, parameters and measurements}
\usage{
get_signal_to_noise_infos(
  regs1,
  regs2,
  beta1,
  beta2,
  Y1,
  Y2,
  names_compare = NULL,
  WIDE = FALSE,
  ROUND = 6
)
}
\arguments{
\item{regs1}{first regressor matrix}

\item{regs2}{second regressor matrix}

\item{beta1}{first parameter vector}

\item{beta2}{second parameter vector}

\item{Y1}{first measurement component}

\item{Y2}{second measurement component}

\item{names_compare}{a character string of two: names for comparisons}

\item{WIDE}{logical; if \code{TRUE}, returns a \code{2 x 7} summary matrix,
else a \code{7 x 2} summary matrix}

\item{ROUND}{a numeric integer giving the digits to round the matrix outputs}
}
\value{
a summary matrix of measures as outlined in the "Description"
}
\description{
Under (P)MCMC sampling, convergence may deteriorate if regressors are higly
colinear, parameters are a-posteriori highly correlated or if the signal to
noise ratio is too high/low. This function helps to diagnose such problems by
printing:
\itemize{
\item{regressor means: }{colmeans of regressor matrix}
\item{regressor variance: }{col-variance of regressor matrix}
\item{beta values: }{true parameter values for the betas}
\item{mean of reg \% * \% beta: }{mean of this vector (matrix product)}
\item{variance of reg \% * \% beta: }{variance of this vector (matrix product)}
\item{mean of measurments Y: }{-}
\item{variance of measurments Y: }{-}
}
}
