% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_pgas_output_transformation.R
\name{pgas_out_2_diagnostics}
\alias{pgas_out_2_diagnostics}
\title{Transforms pgas output from \code{pgas_cpp()} or \code{pgas_R()} to a format
suitable for the function
\code{pmcmcDiagnostics::analyse_mcmc_convergence2()}}
\usage{
pgas_out_2_diagnostics(
  pgas_out,
  par_inits,
  par_names = NULL,
  par_trues = NULL,
  state_names = NULL,
  energy_type_names = NULL,
  NN,
  TT,
  DD
)
}
\arguments{
\item{pgas_out}{output from  \code{pgas_cpp()} or \code{pgas_R()}: a list of
four elements:
\itemize{
\item{\code{sig_sq_xa:}}{matrix of dimension DD x num_mncmc_draws} of
simulated standard deviation values
\item{\code{phi_xa:}}{matrix of dimension DD x num_mncmc_draws} of
simulated autoregressive parameter values
\item{\code{bet_xa}}{matrix of dimension num_par_beta x num_mncmc_draws}
of simulated regressor coefficient values in the order they appear in the
latent state transition equation from d=1,...,DD
}}

\item{par_inits}{list of initialization values for parameter s}

\item{par_names}{list of 3: first component is the name for sigma_squared and
phi, the second component is a list of regressor names for Z-type
regressors, and the last is for U-type regressors}

\item{par_trues}{if a simulation study, then one can provide true parameter
values; the input argument should have the same structure as the argument
\code{par_inits}}

\item{state_names}{character vector of dimension NN giving the names of
the US states}

\item{energy_type_names}{character vector of dimension DD giving the names of
the energy types}

\item{NN}{number of cross sectional units}

\item{TT}{number of time periods}

\item{DD}{dimension of the measurements (latent states)}
}
\value{
a list of 6 elements:
\itemize{
\item{\code{$mcmc_sims:}}{matrix of dimension DD x num_mncmc_draws} of
simulated standard deviation values
\item{\code{$states:}}{matrix of dimension DD x num_mncmc_draws} of
simulated autoregressive parameter values
\item{\code{$par_names}}{matrix of dimension \code{num_par_beta x}
\code{ num_mncmc_draws} of simulated regressor coefficient values in the
order they appear in the latent state transition equation from
d=1,...,DD}
}
}
\description{
Transforms pgas output from \code{pgas_cpp()} or \code{pgas_R()} to a format
suitable for the function
\code{pmcmcDiagnostics::analyse_mcmc_convergence2()}
}
