% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{verify_CIs}
\alias{verify_CIs}
\title{Checks whether HPD is smaller than CI (for the same significance)}
\usage{
verify_CIs(CI, HPD)
}
\arguments{
\item{CI}{a confidence interval as returned via \link{compute_ci}}

\item{HPD}{a confidence interval as returned via \link{compute_hpd}}
}
\value{
side effect: message indicating when HPD is not smaller than CI
}
\description{
Ideally, the HPD is the smallest/densest region around the posterior mean,
so there are computational issues whenever this is not the case (and, here
specifically, whenever the HPD is not smaller than the confidence interval
at the same significance level).
}
