% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{view_diagnostics_table}
\alias{view_diagnostics_table}
\title{Call \code{View} on \code{data.frame} of MCMC convergence diagnostics}
\usage{
view_diagnostics_table(summary_diagnostics, precision, table_name)
}
\arguments{
\item{summary_diagnostics}{\code{data.frame} of MCMC convergence diagnostics}

\item{precision}{digits used for rounding values}

\item{table_name}{name of the table}
}
\value{
pure side effect call to \code{View}
}
\description{
Call \code{View} on \code{data.frame} of MCMC convergence diagnostics
}
