% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table_diagnostics.R
\name{write_diagnostics_table}
\alias{write_diagnostics_table}
\title{Write \code{data.frame} of convergence diagnostics to '.csv'}
\usage{
write_diagnostics_table(summary_diagnostics, par_names, table_path, table_name)
}
\arguments{
\item{summary_diagnostics}{\code{data.frame} of MCMC convergence diagnostics}

\item{par_names}{parameter names}

\item{table_path}{path to write to}

\item{table_name}{name of the table}
}
\value{
pure side effect; writing summary table to \code{.csv} file
}
\description{
Write \code{data.frame} of convergence diagnostics to '.csv'
}
