% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_update_rate_diagnostics.R
\name{get_update_rates}
\alias{get_update_rates}
\title{Generates update rate plots}
\usage{
get_update_rates(states, settings_urs, settings_plots)
}
\arguments{
\item{states}{latent state values}

\item{settings_urs}{settings used for viewing/saving update rate plots}
}
\value{
side effect function; plotting and/or saving
}
\description{
If latent states are provided only!
}
